feat: delete home
