+++ src/routes/+page.js
@@ -1,38 +1,40 @@
 import { postService } from '$lib/supabase.js';
-import { isErrorType } from '$lib/utils/error.js';
+import { isErrorType, handleDatabaseError } from '$lib/utils/error.js';
 
-// 서버 사이드에서 데이터 로드
-export async function load() {
-	try {
-		// 최신 포스트와 추천 포스트를 병렬로 가져오기
-		const [recentPosts, featuredPosts] = await Promise.all([
-			postService.getPosts(4, 0),
-			postService.getFeaturedPosts(2)
-		]);
-
-		return {
-			recentPosts,
-			featuredPosts
-		};
-	} catch (err) {
-		console.error('Error loading posts:', err);
-		
-		// 데이터베이스 테이블이 없는 경우 등은 빈 배열로 처리 (사용자 경험 우선)
-		if (isErrorType(err, 'database')) {
-			const errorMessage = err instanceof Error ? err.message : String(err);
-			if (errorMessage.includes('relation') || errorMessage.includes('does not exist')) {
-				return {
-					recentPosts: [],
-					featuredPosts: []
-				};
-			}
-		}
-		
-		// 기타 심각한 에러는 에러 페이지로 리다이렉트
-		// 하지만 메인 페이지는 사용자 경험을 위해 가능한 한 빈 배열로 처리
-		return {
-			recentPosts: [],
-			featuredPosts: []
-		};
+export async function load({ url }) {
+  const page = parseInt(url.searchParams.get('page') || '1');
+  const limit = 10;
+  const offset = (page - 1) * limit;
+  
+  try {
+    const [posts, allTags] = await Promise.all([
+      postService.getPosts(limit, offset),
+      postService.getAllTags()
+    ]);
+    
+    return {
+      posts,
+      allTags,
+      currentPage: page,
+      hasMore: posts.length === limit
+    };
+  } catch (err) {
+    console.error('Error loading posts:', err);
+    
+    // 데이터베이스 테이블이 없는 경우 등은 빈 배열로 처리
+    if (isErrorType(err, 'database')) {
+      const errorMessage = err instanceof Error ? err.message : String(err);
+      if (errorMessage.includes('relation') || errorMessage.includes('does not exist')) {
+        return {
+          posts: [],
+          allTags: [],
+          currentPage: 1,
+          hasMore: false
+        };
+      }
+    }
+    
+    // 기타 데이터베이스 에러는 에러 페이지로 리다이렉트
+    handleDatabaseError(err, '개발 로그 데이터 불러오기');
 	}
-} 
+}
